* убран batchSize за неиспользованием * int64 в int32

+++ internal/model/folder.go
@@ -1,12 +1,12 @@
 package model
 
 type Folder struct {
-	Id        int64  `db:"id"`
-	IdUser    int64  `db:"id_user"`
-	IdParent  int64  `db:"id_parent"`
-	Lft       int64  `db:"lft"`
-	Rgt       int64  `db:"rgt"`
-	Depth     int64  `db:"depth"`
+	Id        int32  `db:"id"`
+	IdUser    int32  `db:"id_user"`
+	IdParent  int32  `db:"id_parent"`
+	Lft       int32  `db:"lft"`
+	Rgt       int32  `db:"rgt"`
+	Depth     int32  `db:"depth"`
 	Title     string `db:"title"`
 	IsActive  bool   `db:"is_active"`
 	IsProject bool   `db:"is_project"`
